chore: move character stats to independent file
(cherry picked from commit c7c1cf13612cc1fb698f93de0806d10f826878f2)
