+++ Source/wunthshin/Data/Characters/CharacterTableRow/CharacterTableRow.cpp
